apps/backend/services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.26.0
pandas==2.1.4
numpy==1.26.3
